Move TransactionViewItemFactory into Transaction module
