+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/transactions/TransactionsModule.kt
@@ -4,13 +4,11 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.AdapterState
 import io.horizontalsystems.bankwallet.core.App
-import io.horizontalsystems.bankwallet.core.factories.TransactionViewItemFactory
 import io.horizontalsystems.bankwallet.entities.Account
 import io.horizontalsystems.bankwallet.entities.CoinSettings
 import io.horizontalsystems.bankwallet.entities.LastBlockInfo
 import io.horizontalsystems.bankwallet.entities.Wallet
 import io.horizontalsystems.bankwallet.entities.transactionrecords.TransactionRecord
-import io.horizontalsystems.bankwallet.entities.transactionrecords.bitcoin.TransactionLockState
 import io.horizontalsystems.bankwallet.modules.transactionInfo.TransactionInfoAddressMapper
 import io.horizontalsystems.coinkit.models.Coin
 import io.horizontalsystems.core.entities.Currency
@@ -57,51 +55,7 @@ data class TransactionViewItem(
     fun clearRates() {
         mainAmountCurrencyString = null
     }
-
-    sealed class TransactionType {
-        class Incoming(
-            val from: String?,
-            val amount: String,
-            val lockState: TransactionLockState?,
-            val conflictingTxHash: String?
-        ) : TransactionType()
-
-        class Outgoing(
-            val to: String?,
-            val amount: String,
-            val lockState: TransactionLockState?,
-            val conflictingTxHash: String?,
-            val sentToSelf: Boolean
-        ) : TransactionType()
-
-        class Approve(val spender: String, val amount: String, val isMaxAmount: Boolean) :
-            TransactionType()
-
-        class Swap(
-            val exchangeAddress: String,
-            val amountIn: String,
-            val amountOut: String?,
-            val foreignRecipient: Boolean
-        ) : TransactionType()
-
-        class ContractCall(val contractAddress: String, val method: String?) : TransactionType()
-        object ContractCreation : TransactionType()
-
-        override fun equals(other: Any?): Boolean {
-            return when {
-                other is Incoming && this is Incoming -> other.lockState == this.lockState && other.conflictingTxHash == this.conflictingTxHash
-                other is Outgoing && this is Outgoing -> other.lockState == this.lockState && other.conflictingTxHash == this.conflictingTxHash
-                other is Approve && this is Approve -> true
-                other is Swap && this is Swap -> true
-                other is ContractCall && this is ContractCall -> true
-                other is ContractCreation && this is ContractCreation -> true
-                else -> false
-            }
-        }
-    }
-
 }
-
 
 data class TransactionLockInfo(
     val lockedUntil: Date,
